backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.12.0,<3
pydantic-settings==2.1.0
anyio>=3.7.1,<4.0.0

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0
chardet==5.2.0
trafilatura==1.6.3

# ML & Embeddings
openai>=2.7.0
httpx>=0.27.0,<0.28.0
tiktoken==0.5.2

# Vector DB
pinecone-client==3.0.0

# External enrichment
exa-py==1.0.9
requests==2.31.0

# Utilities
aiofiles==23.2.1
python-dotenv==1.0.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1  # MongoDB sync driver
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing

# AWS S3
boto3==1.35.0  # AWS SDK for Python
botocore==1.35.0  # Low-level AWS service access
